updating changes

+++ src/info/guardianproject/otr/app/im/app/adapter/ChatListenerAdapter.java
@@ -1,13 +1,13 @@
 /*
  * Copyright (C) 2007-2008 Esmertec AG. Copyright (C) 2007-2008 The Android Open
  * Source Project
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
  * the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
@@ -17,15 +17,13 @@
 
 package info.guardianproject.otr.app.im.app.adapter;
 
+import info.guardianproject.otr.app.im.IChatListener;
+import info.guardianproject.otr.app.im.IChatSession;
 import info.guardianproject.otr.app.im.app.ImApp;
 import info.guardianproject.otr.app.im.engine.Contact;
 import info.guardianproject.otr.app.im.engine.ImErrorInfo;
 import info.guardianproject.otr.app.im.engine.Message;
-
-import info.guardianproject.otr.app.im.IChatSession;
-import info.guardianproject.otr.app.im.IChatListener;
 import info.guardianproject.util.LogCleaner;
-
 import android.os.RemoteException;
 import android.util.Log;
 
@@ -49,7 +47,7 @@ public class ChatListenerAdapter extends IChatListener.Stub {
         if (Log.isLoggable(TAG, Log.DEBUG)) {
             LogCleaner.debug(TAG, "onIncomingMessage(" + ses + ", " + msg + ")");
         }
-        
+
         return true;
     }
 
@@ -99,7 +97,7 @@ public class ChatListenerAdapter extends IChatListener.Stub {
     }
 
     @Override
-    public void onIncomingFileTransferProgress(String file, float percent) throws RemoteException {
+    public void onIncomingFileTransferProgress(String file, int percent) throws RemoteException {
         if (Log.isLoggable(TAG, Log.DEBUG)) {
             LogCleaner.debug(TAG, "onIncomingFileTransferProgress(" + file + "," + percent + ")");
         }
@@ -110,6 +108,7 @@ public class ChatListenerAdapter extends IChatListener.Stub {
         if (Log.isLoggable(TAG, Log.DEBUG)) {
             LogCleaner.debug(TAG, "onIncomingFileTransferError(" + file + "," + message + ")");
         }
-        
+
     }
+
 }
